Moved ObjectHandle to a non plugin directory
