+++ src/engine/plugins/three/handles/BallHandle.ts
@@ -1,4 +1,4 @@
-import { ObjectHandle } from "./ObjectHandle";
+import { ObjectHandle } from "../../../ObjectHandle";
 import { SimBall } from "../objects/SimBall";
 
 export class BallHandle extends ObjectHandle<SimBall> {
