+++ src/engine/plugins/three/handles/BoxHandle.ts
@@ -1,4 +1,4 @@
-import { ObjectHandle } from "./ObjectHandle";
+import { ObjectHandle } from "../../../ObjectHandle";
 import { SimBox } from "../objects/SimBox";
 
 export class BoxHandle extends ObjectHandle<SimBox> {
