+++ src/engine/plugins/three/handles/PyramidHandle.ts
@@ -1,4 +1,4 @@
-import { ObjectHandle } from "./ObjectHandle";
+import { ObjectHandle } from "../../../ObjectHandle";
 import { SimPyramid } from "../objects/SimPyramid";
 
 export class PyramidHandle extends ObjectHandle<SimPyramid> {
