+++ src/engine/plugins/three/handles/WallHandle.ts
@@ -1,4 +1,4 @@
-import { ObjectHandle } from "./ObjectHandle";
+import { ObjectHandle } from "../../../ObjectHandle";
 import { SimWall } from "../objects/SimWall";
 
 export class WallHandle extends ObjectHandle<SimWall> {}
